requirements.txt
asgiref==3.8.1
Django==5.1.6
sqlparse==0.5.3
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
requests==2.32.3
urllib3==2.3.0
djangorestframework==3.15.2
cffi==1.17.1
cryptography==44.0.1
pycparser==2.22
